server/app/requirements.txt
tensorflow==2.6.0
keras==2.6.0
tensorflow_hub==0.11.0
tensorflow-estimator==2.6.0
uvicorn[standard]
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
gast==0.4.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
libclang==12.0.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
PyMySQL==1.0.2
python-multipart
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
termcolor==1.1.0
matplotlib
opencv-python
mysqlclient
